requirements.txt
# FastAPI and Web Server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# File Handling
python-multipart==0.0.19
pdfplumber==0.11.5
PyPDF2==3.0.1
Pillow==11.0.0
pdf2image==1.17.0

# HTTP Client for NVIDIA NIM API
httpx==0.28.1

# Environment Variables
python-dotenv==1.0.1

# Additional Utilities
typing-extensions==4.12.2

